Create separate API folder
